refactor: 등록이미지 컴포넌트 분리
